refactor(plugin-react-query): move operation dispatcher to @internals/tanstack-query
Move the query/mutation classify-and-dispatch generator into
@internals/tanstack-query as createOperationDispatcher, next to the
classifyOperation it's built on, instead of duplicating it inline in
plugin-react-query's plugin.ts. Other TanStack-family plugins (vue-query,
swr) have the same query/mutation config shape and can reuse it later.

plugin.ts now just builds the query and mutation generator lists and
passes them to the shared factory.
