+++ internals/tanstack-query/src/index.ts
@@ -1,5 +1,6 @@
 export { InfiniteQueryOptions } from './components/InfiniteQueryOptions.tsx'
 export { MutationKey, mutationKeyTransformer } from './components/MutationKey.tsx'
+export { createOperationDispatcher } from './operationDispatcher.ts'
 export { createMutationResolver, createQueryResolver, type MutationResolver, type QueryResolver, type QueryVariant } from './resolver.ts'
 export { QueryKey, queryKeyTransformer } from './components/QueryKey.tsx'
 export type { Infinite, Mutation, ParamsCasing, ParamsType, PathParamsType, Query, Transformer } from './types.ts'
